requirements.txt
aiogram
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
celery
redis
python-dotenv
pydantic
pydantic-settings
aiohttp
